moved ManualVelocitySmoothing to motion_planning lib
