+++ src/lib/motion_planning/ManualVelocitySmoothingXY.hpp
